Share more code from TestCasesRunner directory
