+++ src/tools/illink/test/Mono.Linker.Tests.Shared/BaseMetadataProvider.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+#nullable enable
 using System;
 using System.Linq;
 using Mono.Cecil;
