+++ src/tools/illink/test/Mono.Linker.Tests.Shared/FormattingUtils.cs
