+++ src/tools/illink/test/Mono.Linker.Tests.Shared/ManagedCompilationResult.cs
