+++ src/tools/illink/test/Mono.Linker.Tests.Shared/ObjectFactory.cs
