+++ src/tools/illink/test/Mono.Linker.Tests.Shared/PathUtilities.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+#nullable enable
 using System;
 using System.IO;
 using System.Runtime.CompilerServices;
@@ -25,7 +26,7 @@ namespace Mono.Linker.Tests.TestCasesRunner
 #error "Unknown TFM"
 #endif
 
-		public static string GetTestsSourceRootDirectory ([CallerFilePath] string thisFile = null) =>
+		public static string GetTestsSourceRootDirectory ([CallerFilePath] string? thisFile = null) =>
 			Path.GetFullPath ((string) AppContext.GetData ("Mono.Linker.Tests.LinkerTestDir")!);
 
 		public static string GetTestAssemblyRoot (string assemblyName)
