+++ src/tools/illink/test/Mono.Linker.Tests.Shared/SetupCompileInfo.cs
@@ -14,7 +14,7 @@ namespace Mono.Linker.Tests.TestCasesRunner
 		public string[] Defines;
 		public string[] References;
 		public SourceAndDestinationPair[] Resources;
-		public string AdditionalArguments;
+		public string[] AdditionalArguments;
 		public string CompilerToUse;
 		public bool AddAsReference;
 		public bool RemoveFromLinkerInput;
