--- src/tools/illink/test/Mono.Linker.Tests/TestCasesRunner/SourceAndDestinationPair.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using Mono.Linker.Tests.Extensions;
-
-namespace Mono.Linker.Tests.TestCasesRunner
-{
-	public class SourceAndDestinationPair
-	{
-		public NPath Source;
-		public string DestinationFileName;
-	}
-}
